--- starter-code/src/app/app.component.html
@@ -1,3 +1,0 @@
-<h1>
-  {{title}}
-</h1>
