+++ starters-Code/src/app/app.component.html
@@ -1,0 +1,6 @@
+<nav>
+  <a [routerLink]="['home']">Home</a>
+
+</nav>
+<router-outlet></router-outlet>
+
