+++ starters-Code/src/app/movie.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { MovieService } from './movie.service';
+
+describe('MovieService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [MovieService]
+    });
+  });
+
+  it('should be created', inject([MovieService], (service: MovieService) => {
+    expect(service).toBeTruthy();
+  }));
+});
