+++ starters-Code/src/app/my-movie-component/my-movie-component.component.html
@@ -1,0 +1,2 @@
+Este es el {{id}}
+
